==> 2. GCS.ipynb <==
# This is the 2nd notebook
- To collect GCS data (latitude and longitude) for each station

# Import Necessary Library

# data maniuplation
import pandas as pd
# web scraping
import requests
from bs4 import BeautifulSoup

# Web Scrape

# This is the link include the region and its respective latitude and longitude
link = "https://www.hko.gov.hk/en/cis/stn.htm"
response = requests.get(link)

# Parse html

# parse html
soup = BeautifulSoup(response.text, 'html.parser')

# get the list of region 
region = [tag.get_text(separator="\n").split("\n")[0].strip("*") for tag in soup.select("td:nth-of-type(1)")[1:]]

# get the list of latitude
latitude = [lat.text for lat in soup.select("td:nth-of-type(2)")]

# get the list of longitude
longitude = [long.text for long in soup.select("td:nth-of-type(3)")]

# Store in DataFrame

# combine region, latitude and longitude into dictionary
dict_gcs = {'region': region, 'latitude_raw': latitude, 'longitude_raw': longitude}

# create dataframe from dictionary
df_gcs = pd.DataFrame(dict_gcs).drop_duplicates().reset_index(drop=True).iloc[:80]

# Angle Conversion

def angle_conversion(angle):
    '''
    Output: angle in degree with decimal
    Input: angle in degree, minute and second
    Note: identify degree, minute and second by its sign 
    '''
    degree_sign = "°"
    minute_sign = "'"
    second_sign = '"'
    all_sign = degree_sign + minute_sign + second_sign
    if degree_sign in angle:
        degree = float(angle.split(degree_sign)[0].strip(all_sign))
    else:
        degree = 0
    if minute_sign in angle:
        minute = float(angle.split(minute_sign)[0][-2:].strip(all_sign))
    else:
        minute = 0
    if second_sign in angle:
        second = float(angle.split(second_sign)[0][-2:].strip(all_sign))
    else:
        second = 0
    degree_total = degree + minute / 60 + second / 3600
    return degree_total

# convert the latitude to degree only
df_gcs['latitude'] = df_gcs['latitude_raw'].apply(angle_conversion)

# convert the longitude to degree only
df_gcs['longitude'] = df_gcs['longitude_raw'].apply(angle_conversion)

# include desired columns only
df_gcs_final = df_gcs[['region', 'latitude', 'longitude']]

df_gcs_final

# Save as csv

# save it in csv
df_gcs_final.to_csv("gcs.csv")
