==> 3. Combine.ipynb <==
# This is the 3rd notebook
- To combine weather data and GCS data

# Import Necessary Library

# data maniuplation
import pandas as pd

# Read CSV of Data and GCS (Geographic Coordinate System)

# read csv for data series and gcs, ignore addition index column by setting index_col = 0
data = pd.read_csv("data_final.csv", index_col=0)
gcs = pd.read_csv("gcs.csv", index_col=0)

# merge two dataframe by region as key
combine = data.merge(gcs, on='region', how='left')

combine

# Save as csv

# save it in csv
combine.to_csv("data_gcs_combine.csv")
