==> 4. Region_update.ipynb <==
# This is the 4th notebook
- To move value from old region to new region

# Import Necessary Library

# data maniuplation
import pandas as pd

# read csv that combined data series and gcs, ignore addition index column by setting index_col = 0
combine = pd.read_csv("data_gcs_combine.csv", index_col=0)

# Fill Null for New Region from Old Region

def region_update(df, old, new):
    '''
    Output: dataframe without old region
    Input: dataframe, old region, new region
    Note: remove old region after filling null for new region by old region
    '''
    df = df.copy()
    df_new = df[df['region'] == new]
    df_old = df[df['region'] == old]
    df_new_reset_index = df_new.reset_index()
    df_old_reset_index = df_old.reset_index()
    update = df_new_reset_index.combine_first(df_old_reset_index).set_index('index')
    df_new_index = df_new.index.tolist()
    df_old_index = df_old.index.tolist()
    list_drop = df_new_index + df_old_index
    df_drop = df.drop(list_drop)
    df_combine = pd.concat([df_drop, update])
    df_sort = df_combine.sort_values(['date','region']).reset_index(drop=True)
    return df_sort

# define the pair of new region and old region
region_old = "Tai Po(Conservation Studies Centre)"
region_new = "Tai Po(Yuen Chau Tsai Park)"

# apply the pair of new region and old region to combined dataframe
combine_update = region_update(combine, region_old, region_new)

combine_update

# Save as csv

# save it in csv
combine_update.to_csv("data_gcs_combine_update.csv")
