==> 5. District_update.ipynb <==
# This is the 5th notebook
- To map region to district

# Import Necessary Library

# data maniuplation
import pandas as pd

# Read CSV

# read csv that updated region, ignore addition index column by setting index_col = 0
df = pd.read_csv("data_gcs_combine_update.csv", index_col=0)

# Create a map for Station to District

# define a dictionary to map 51 regions to 18 districts in Hong Kong
station_to_district = {
    "Bluff Head": "Southern",
    "Central Pier": "Central and Western",
    "Cheung Chau": "Islands",
    "Cheung Chau Beach": "Islands",
    "Ching Pak House(Tsing Yi)": "Kwai Tsing",
    "Clear Water Bay": "Sai Kung",
    "Green Island": "Central and Western",
    "Happy Valley": "Wan Chai",
    "Hong Kong International Airport": "Islands",
    "Hong Kong Observatory": "Yau Tsim Mong",
    "Hong Kong Park": "Central and Western",
    "Kai Tak": "Kowloon City",
    "Kai Tak Runway Park": "Kowloon City",
    "Kau Sai Chau": "Sai Kung",
    "King's Park": "Yau Tsim Mong",
    "Kowloon City": "Kowloon City",
    "Kwun Tong": "Kwun Tong",
    "Lamma Island": "Islands",
    "Lau Fau Shan": "Yuen Long",
    "New Tsing Yi Station": "Kwai Tsing",
    "Ngong Ping": "Islands",
    "North Point": "Eastern",
    "Pak Tam Chung (Tsak Yue Wu)": "Sai Kung",
    "Peng Chau": "Islands",
    "Sai Kung": "Sai Kung",
    "Sha Chau": "Tuen Mun",
    "Sha Tin": "Sha Tin",
    "Sham Shui Po": "Sham Shui Po",
    "Shau Kei Wan": "Eastern",
    "Shek Kong": "Yuen Long",
    "Shell Oil Depot": "Kwai Tsing",
    "Sheung Shui": "North",
    "Stanley": "Southern",
    "Star Ferry(Kowloon)": "Yau Tsim Mong",
    "Ta Kwu Ling": "North",
    "Tai Mei Tuk": "Tai Po",
    "Tai Mo Shan": "Tsuen Wan",
    "Tai Po Kau": "Tai Po",
    "Tai Po(Yuen Chau Tsai Park)": "Tai Po",
    "Tap Mun East": "Tai Po",
    "Tate's Cairn": "Sha Tin",
    "The Peak": "Central and Western",
    "Tseung Kwan O": "Sai Kung",
    "Tsuen Wan": "Tsuen Wan",
    "Tsuen Wan Shing Mun Valley": "Tsuen Wan",
    "Tuen Mun Children and Juvenile Home": "Tuen Mun",
    "Tuen Mun Government Offices": "Tuen Mun",
    "Waglan Island": "Islands",
    "Wetland Park": "Yuen Long",
    "Wong Chuk Hang": "Southern",
    "Wong Tai Sin": "Wong Tai Sin"
}

# Mapping

# map by dictionary
df['district'] = df['region'].map(station_to_district)

# Save as csv

# save it in csv
df.to_csv("data_gcs_combine_update_district.csv")
